Reorganizing grouping file implementation
