+++ tt_metal/fabric/physical_grouping_descriptor_matching.cpp
@@ -1,0 +1,669 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <stdexcept>
+#include <fstream>
+#include <sstream>
+#include <ostream>
+#include <filesystem>
+#include <algorithm>
+#include <unordered_set>
+#include <unordered_map>
+#include <set>
+#include <queue>
+#include <memory>
+#include <cctype>
+#include <functional>
+#include <optional>
+#include <tt_stl/assert.hpp>
+#include <fmt/format.h>
+
+#include "protobuf/physical_grouping_descriptor.pb.h"
+#include "protobuf/mesh_graph_descriptor.pb.h"
+#include <tt-metalium/experimental/fabric/physical_grouping_descriptor.hpp>
+#include <tt-metalium/experimental/fabric/mesh_graph_descriptor.hpp>
+#include <tt-metalium/experimental/fabric/topology_solver.hpp>
+#include <tt-metalium/experimental/fabric/topology_mapper_utils.hpp>
+#include "tt_metal/fabric/physical_system_descriptor.hpp"
+#include <tt-logger/tt-logger.hpp>
+#include <cctype>
+#include <map>
+
+#include <google/protobuf/text_format.h>
+
+using namespace tt::tt_fabric;
+
+namespace {
+
+// Forward declarations for MGD helper functions (defined in graph_building.cpp)
+AdjacencyGraph<uint32_t> build_mgd_mesh_instance_adjacency(
+    const MeshGraphDescriptor& mesh_graph_descriptor, GlobalNodeId mesh_instance_id);
+AdjacencyGraph<uint32_t> build_mgd_graph_instance_adjacency(
+    const MeshGraphDescriptor& mesh_graph_descriptor, GlobalNodeId graph_instance_id);
+
+// Convert MGD instances to GroupingInfo map (includes adjacency graphs and ASIC counts)
+// Calculates required ASIC counts bottom-up and builds adjacency graphs
+// Returns map: (type, name) -> GroupingInfo
+std::unordered_map<std::string, std::unordered_map<std::string, GroupingInfo>> build_mgd_to_grouping_info_map(
+    const MeshGraphDescriptor& mesh_graph_descriptor) {
+    std::unordered_map<std::string, std::unordered_map<std::string, GroupingInfo>> mgd_grouping_infos;
+
+    // ===== Step 1: Calculate required ASIC counts bottom-up =====
+    // Map: (type, name) -> required_asics
+    std::unordered_map<std::string, std::unordered_map<std::string, uint32_t>> required_asics_map;
+
+    // Step 1a: Calculate required ASICs for all mesh instances (bottom level)
+    for (GlobalNodeId mesh_id : mesh_graph_descriptor.all_meshes()) {
+        const auto& mesh_instance = mesh_graph_descriptor.get_instance(mesh_id);
+        uint32_t required_chips = mesh_graph_descriptor.get_chip_count(mesh_id);
+        required_asics_map[mesh_instance.type][mesh_instance.name] = required_chips;
+    }
+
+    // Step 1b: Calculate required ASICs for graph instances bottom-up (children before parents)
+    // Process graphs in topological order by iterating until all are processed
+    std::unordered_set<GlobalNodeId> processed_graphs;
+    bool progress_made = true;
+
+    while (progress_made) {
+        progress_made = false;
+
+        for (GlobalNodeId graph_id : mesh_graph_descriptor.all_graphs()) {
+            if (processed_graphs.contains(graph_id)) {
+                continue;  // Already processed
+            }
+
+            const auto& graph_instance = mesh_graph_descriptor.get_instance(graph_id);
+            const std::string& graph_type = graph_instance.type;
+            const std::string& graph_name = graph_instance.name;
+
+            // Check if all sub-instances have been processed (have required_asics calculated)
+            bool all_sub_instances_ready = true;
+            uint32_t required_asics = 0;
+
+            for (GlobalNodeId sub_id : graph_instance.sub_instances) {
+                const auto& sub_instance = mesh_graph_descriptor.get_instance(sub_id);
+
+                // Check if this sub-instance's required_asics is already calculated
+                auto sub_type_it = required_asics_map.find(sub_instance.type);
+                if (sub_type_it == required_asics_map.end()) {
+                    all_sub_instances_ready = false;
+                    break;
+                }
+
+                auto sub_name_it = sub_type_it->second.find(sub_instance.name);
+                if (sub_name_it == sub_type_it->second.end()) {
+                    all_sub_instances_ready = false;
+                    break;
+                }
+
+                required_asics += sub_name_it->second;
+            }
+
+            // If all sub-instances are ready, calculate and store this graph's required_asics
+            if (all_sub_instances_ready) {
+                required_asics_map[graph_type][graph_name] = required_asics;
+                processed_graphs.insert(graph_id);
+                progress_made = true;
+            }
+        }
+    }
+
+    // Verify all graphs were processed (should not have cycles, but check for safety)
+    for (GlobalNodeId graph_id : mesh_graph_descriptor.all_graphs()) {
+        const auto& graph_instance = mesh_graph_descriptor.get_instance(graph_id);
+        auto type_it = required_asics_map.find(graph_instance.type);
+        if (type_it == required_asics_map.end() || !type_it->second.contains(graph_instance.name)) {
+            TT_THROW(
+                "Failed to calculate required ASIC count for graph instance '{}' (type '{}'). "
+                "This may indicate a circular dependency in the MGD.",
+                graph_instance.name,
+                graph_instance.type);
+        }
+    }
+
+    // ===== Step 2: Build GroupingInfo objects with adjacency graphs and ASIC counts =====
+
+    // Process mesh instances
+    // Store only one entry per mesh definition name (M0, M1), not per instance (M0_0, M0_1, etc.)
+    std::set<std::string> processed_mesh_definitions;
+    for (GlobalNodeId mesh_id : mesh_graph_descriptor.all_meshes()) {
+        const auto& mesh_instance = mesh_graph_descriptor.get_instance(mesh_id);
+        const std::string& mesh_type = mesh_instance.type;
+        const std::string& mesh_name = mesh_instance.name;
+
+        // Skip if we've already processed this mesh definition
+        if (processed_mesh_definitions.contains(mesh_name)) {
+            continue;
+        }
+        processed_mesh_definitions.insert(mesh_name);
+
+        // Build adjacency graph for this mesh instance (use first instance of this mesh definition)
+        AdjacencyGraph<uint32_t> adjacency_graph = build_mgd_mesh_instance_adjacency(mesh_graph_descriptor, mesh_id);
+
+        // Get required ASIC count (calculated above)
+        uint32_t asic_count = required_asics_map.at(mesh_type).at(mesh_name);
+
+        // Get device topology dimensions for corner orientation assignment
+        const auto* mesh_desc = std::get<const proto::MeshDescriptor*>(mesh_instance.desc);
+        TT_FATAL(mesh_desc != nullptr, "Mesh descriptor is null");
+        const auto& device_topology = mesh_desc->device_topology();
+        std::vector<int32_t> device_dims(device_topology.dims().begin(), device_topology.dims().end());
+
+        // Create GroupingInfo
+        GroupingInfo grouping_info;
+        grouping_info.name = mesh_name;  // Keep original name for matching
+        grouping_info.type = mesh_type;
+        grouping_info.asic_count = asic_count;
+        grouping_info.adjacency_graph = std::move(adjacency_graph);
+
+        // Create a single item representing the mesh (for corner orientation assignment)
+        // The item represents the entire mesh as a single unit
+        GroupingItemInfo mesh_item;
+        mesh_item.type = GroupingItemInfo::ItemType::GROUPING_REF;
+        mesh_item.grouping_name = mesh_name;
+        grouping_info.items.push_back(std::move(mesh_item));
+
+        // Assign corner orientations based on mesh dimensions
+        // For mesh instances with a single item, the helper function will assign corners appropriately
+        assign_corner_orientations_to_grouping(grouping_info, device_dims);
+
+        // Store keyed by mesh definition name (not instance key)
+        mgd_grouping_infos[mesh_type][mesh_name] = std::move(grouping_info);
+    }
+
+    // Process graph instances
+    for (GlobalNodeId graph_id : mesh_graph_descriptor.all_graphs()) {
+        const auto& graph_instance = mesh_graph_descriptor.get_instance(graph_id);
+        const std::string& graph_type = graph_instance.type;
+        const std::string& graph_name = graph_instance.name;
+
+        // Skip if already processed (same name/type)
+        if (mgd_grouping_infos.contains(graph_type) && mgd_grouping_infos.at(graph_type).contains(graph_name)) {
+            continue;
+        }
+
+        // Build adjacency graph for this graph instance
+        AdjacencyGraph<uint32_t> adjacency_graph = build_mgd_graph_instance_adjacency(mesh_graph_descriptor, graph_id);
+
+        // Get required ASIC count (calculated above)
+        uint32_t asic_count = required_asics_map.at(graph_type).at(graph_name);
+
+        // Create GroupingInfo
+        GroupingInfo grouping_info;
+        grouping_info.name = graph_name;
+        grouping_info.type = graph_type;
+        grouping_info.asic_count = asic_count;
+        grouping_info.adjacency_graph = std::move(adjacency_graph);
+        // items left empty - not needed for matching
+
+        mgd_grouping_infos[graph_type][graph_name] = std::move(grouping_info);
+    }
+
+    return mgd_grouping_infos;
+}
+
+// -----------------------------------------------------------------------------
+// Phase 3: Higher-layer graph matching helpers
+// -----------------------------------------------------------------------------
+
+bool is_mgd_graph_ready(
+    const MeshGraphDescriptor& mesh_graph_descriptor,
+    const std::string& graph_name,
+    const std::unordered_map<std::string, std::unordered_map<std::string, std::vector<GroupingInfo>>>& result,
+    const std::unordered_map<std::string, std::string>& known_mappings) {
+    const auto& instance_ids = mesh_graph_descriptor.instances_by_name(graph_name);
+    if (instance_ids.empty()) {
+        return false;
+    }
+    const auto& graph_instance = mesh_graph_descriptor.get_instance(instance_ids[0]);
+    for (GlobalNodeId sub_id : graph_instance.sub_instances) {
+        const auto& sub_instance = mesh_graph_descriptor.get_instance(sub_id);
+        if (!result.contains(sub_instance.type) && !known_mappings.contains(sub_instance.type)) {
+            return false;
+        }
+    }
+    return true;
+}
+
+bool mgd_graph_depends_on(
+    const MeshGraphDescriptor& mesh_graph_descriptor, const std::string& dep_graph_name, const std::string& on_type) {
+    const auto& instance_ids = mesh_graph_descriptor.instances_by_name(dep_graph_name);
+    if (instance_ids.empty()) {
+        return false;
+    }
+    const auto& graph_instance = mesh_graph_descriptor.get_instance(instance_ids[0]);
+    for (GlobalNodeId sub_id : graph_instance.sub_instances) {
+        const auto& sub_instance = mesh_graph_descriptor.get_instance(sub_id);
+        if (sub_instance.type == on_type) {
+            return true;
+        }
+    }
+    return false;
+}
+
+bool pgd_grouping_depends_on(const GroupingInfo& pgd_grouping, const std::string& on_type) {
+    for (const auto& item : pgd_grouping.items) {
+        if (item.type == GroupingItemInfo::ItemType::GROUPING_REF && item.grouping_name == on_type) {
+            return true;
+        }
+    }
+    return false;
+}
+
+void process_higher_layer_and_recurse(
+    const MeshGraphDescriptor& mesh_graph_descriptor,
+    const std::unordered_map<std::string, std::unordered_map<std::string, GroupingInfo>>& mgd_grouping_infos,
+    const std::unordered_map<std::string, std::vector<GroupingInfo>>& resolved_groupings_cache_,
+    std::unordered_map<std::string, std::unordered_map<std::string, std::vector<GroupingInfo>>>& result,
+    std::unordered_map<std::string, std::string>& known_mappings,
+    const std::string& mgd_type,
+    const std::string& graph_name) {
+    if (result.contains(mgd_type) && result.at(mgd_type).contains(graph_name)) {
+        return;
+    }
+
+    const auto& instance_ids = mesh_graph_descriptor.instances_by_name(graph_name);
+    if (instance_ids.empty()) {
+        return;
+    }
+    GlobalNodeId repr_graph_id = instance_ids[0];
+    if (!is_mgd_graph_ready(mesh_graph_descriptor, graph_name, result, known_mappings)) {
+        return;
+    }
+
+    const auto& graph_instance = mesh_graph_descriptor.get_instance(repr_graph_id);
+
+    std::unordered_set<std::string> allowed_pgd_child_types;
+    for (GlobalNodeId sub_id : graph_instance.sub_instances) {
+        const auto& sub_instance = mesh_graph_descriptor.get_instance(sub_id);
+        auto it = known_mappings.find(sub_instance.type);
+        if (it != known_mappings.end()) {
+            allowed_pgd_child_types.insert(it->second);
+        }
+    }
+
+    AdjacencyGraph<uint32_t> mgd_adjacency = build_mgd_graph_instance_adjacency(mesh_graph_descriptor, repr_graph_id);
+
+    size_t mgd_nodes = mgd_adjacency.get_nodes().size();
+    if (mgd_nodes == 0) {
+        return;
+    }
+
+    std::vector<GroupingInfo> matches;
+    for (const auto& [pgd_type, pgd_groupings] : resolved_groupings_cache_) {
+        if (pgd_type == "MESH") {
+            continue;
+        }
+        for (const auto& pgd_grouping : pgd_groupings) {
+            // PGD grouping must depend on one of the allowed child types
+            bool depends_on_allowed = false;
+            for (const std::string& allowed_type : allowed_pgd_child_types) {
+                if (pgd_grouping_depends_on(pgd_grouping, allowed_type)) {
+                    depends_on_allowed = true;
+                    break;
+                }
+            }
+            if (!depends_on_allowed) {
+                continue;
+            }
+
+            size_t pgd_nodes = pgd_grouping.adjacency_graph.get_nodes().size();
+            if (pgd_nodes < mgd_nodes) {
+                continue;
+            }
+
+            auto mapping_result = solve_topology_mapping<uint32_t, uint32_t>(
+                mgd_adjacency, pgd_grouping.adjacency_graph, {}, ConnectionValidationMode::STRICT, true);
+
+            if (mapping_result.success) {
+                matches.push_back(pgd_grouping);
+            }
+        }
+    }
+
+    if (!matches.empty()) {
+        const GroupingInfo* best = matches.data();
+        for (const auto& m : matches) {
+            if (m.adjacency_graph.get_nodes().size() == mgd_nodes) {
+                best = &m;
+                break;
+            }
+        }
+        result[mgd_type][graph_name].push_back(*best);
+        known_mappings[mgd_type] = best->type;
+    } else {
+        // No matches found - use the MGD grouping info itself
+        auto mgd_it = mgd_grouping_infos.find(mgd_type);
+        if (mgd_it != mgd_grouping_infos.end()) {
+            auto instance_it = mgd_it->second.find(graph_name);
+            if (instance_it != mgd_it->second.end()) {
+                result[mgd_type][graph_name].push_back(instance_it->second);
+            }
+        }
+    }
+
+    for (const auto& [dep_mgd_type, dep_instances] : mgd_grouping_infos) {
+        if (dep_mgd_type == "MESH") {
+            continue;
+        }
+        for (const auto& [dep_graph_name, _] : dep_instances) {
+            if (!mgd_graph_depends_on(mesh_graph_descriptor, dep_graph_name, mgd_type)) {
+                continue;
+            }
+            if (!is_mgd_graph_ready(mesh_graph_descriptor, dep_graph_name, result, known_mappings)) {
+                continue;
+            }
+            if (result.contains(dep_mgd_type) && result.at(dep_mgd_type).contains(dep_graph_name)) {
+                continue;
+            }
+            process_higher_layer_and_recurse(
+                mesh_graph_descriptor,
+                mgd_grouping_infos,
+                resolved_groupings_cache_,
+                result,
+                known_mappings,
+                dep_mgd_type,
+                dep_graph_name);
+        }
+    }
+}
+
+}  // namespace
+
+namespace tt::tt_fabric {
+
+ValidGroupingsMap PhysicalGroupingDescriptor::get_valid_groupings_for_mgd(
+    const MeshGraphDescriptor& mesh_graph_descriptor) const {
+    return get_valid_groupings_for_mgd(
+        mesh_graph_descriptor, static_cast<const tt::tt_metal::PhysicalSystemDescriptor*>(nullptr));
+}
+
+ValidGroupingsMap PhysicalGroupingDescriptor::get_valid_groupings_for_mgd(
+    const MeshGraphDescriptor& mesh_graph_descriptor,
+    const tt::tt_metal::PhysicalSystemDescriptor& physical_system_descriptor) const {
+    return get_valid_groupings_for_mgd(mesh_graph_descriptor, &physical_system_descriptor);
+}
+
+ValidGroupingsMap PhysicalGroupingDescriptor::get_valid_groupings_for_mgd(
+    const MeshGraphDescriptor& mesh_graph_descriptor,
+    const tt::tt_metal::PhysicalSystemDescriptor* physical_system_descriptor) const {
+    ValidGroupingsMap result;
+
+    // ===== PHASE 0: Convert MGD instances to GroupingInfo map (includes adjacency graphs and ASIC counts) =====
+    // This step calculates required ASIC counts bottom-up and builds adjacency graphs
+    std::unordered_map<std::string, std::unordered_map<std::string, GroupingInfo>> mgd_grouping_infos =
+        build_mgd_to_grouping_info_map(mesh_graph_descriptor);
+
+    // ===== PHASE 1: Build flattened adjacency graphs for all mesh group infos =====
+    // Each possibility from build_flattened_adjacency_mesh gets a uniquified key (name_0, name_1, ...)
+    std::unordered_map<std::string, GroupingInfo> mesh_flat_groupings;  // Lookup map for flattened GroupingInfo by key
+    auto mesh_it = resolved_groupings_cache_.find("MESH");
+    if (mesh_it != resolved_groupings_cache_.end()) {
+        for (const auto& mesh_group_info : mesh_it->second) {
+            auto meshes = build_flattened_adjacency_mesh(mesh_group_info, physical_system_descriptor);
+            for (size_t i = 0; i < meshes.size(); ++i) {
+                std::string uniquified_key = mesh_group_info.name + "_" + std::to_string(i);
+                mesh_flat_groupings[uniquified_key] = std::move(meshes[i]);
+            }
+        }
+    } else {
+        TT_THROW("Internal error: MESH grouping not found in resolved_groupings_cache_");
+    }
+
+    // ===== PHASE 2: Match MESH mgd groupings to MESH groupings =====
+    // For each MGD mesh instance, find all valid PGD mesh groupings that can contain it
+    for (const auto& [mgd_instance_key, mgd_mesh_grouping] : mgd_grouping_infos["MESH"]) {
+        const std::string& instance_name = mgd_instance_key;  // Use unique instance key (includes mesh_id)
+        const GroupingInfo& mgd_grouping_info = mgd_mesh_grouping;
+        const std::string& instance_type = mgd_grouping_info.type;  // Should be "MESH"
+
+        // Required nodes from MGD adjacency graph (this represents the topology pattern to match)
+        size_t required_nodes = mgd_grouping_info.adjacency_graph.get_nodes().size();
+
+        // Group valid candidates by node difference (map is ordered by key ascending)
+        std::map<size_t, std::vector<std::string>> candidates_by_diff;
+        for (const auto& [name, grouping_info] : mesh_flat_groupings) {
+            size_t n = grouping_info.adjacency_graph.get_nodes().size();
+            if (n >= required_nodes) {
+                candidates_by_diff[n - required_nodes].push_back(name);
+            }
+        }
+
+        // Process difference levels from closest to farthest; stop at first level with any match
+        std::vector<std::string> best_matches;
+        for (const auto& [node_diff, names] : candidates_by_diff) {
+            (void)node_diff;
+            for (const std::string& name : names) {
+                const auto& grouping_info = mesh_flat_groupings.at(name);
+                // NOTE: If we ever want to support mixed type topologies, we need to add constraints to match the types
+                auto mapping_result = solve_topology_mapping<uint32_t, uint32_t>(
+                    mgd_grouping_info.adjacency_graph,
+                    grouping_info.adjacency_graph,
+                    {},
+                    ConnectionValidationMode::STRICT,
+                    true);
+                if (mapping_result.success) {
+                    best_matches.push_back(name);
+                }
+            }
+            if (!best_matches.empty()) {
+                break;  // Found matches at this (best) level
+            }
+        }
+
+        // Store all best matches (add all entries that are possible)
+        if (best_matches.empty()) {
+            // No match found - use the MGD grouping info itself
+            result[instance_type][instance_name].push_back(mgd_grouping_info);
+        } else {
+            for (const std::string& match_name : best_matches) {
+                // Look up the flattened GroupingInfo from lookup map (already contains flattened adjacency graphs)
+                auto lookup_it = mesh_flat_groupings.find(match_name);
+                if (lookup_it != mesh_flat_groupings.end()) {
+                    // Return the flattened GroupingInfo (not the original)
+                    const GroupingInfo& flattened_grouping = lookup_it->second;
+                    result[instance_type][instance_name].push_back(flattened_grouping);
+                }
+            }
+        }
+    }
+
+    // =============================================================================
+    // Phase 3: Higher-layer graph matching (FABRIC, SUPER_FABRIC, etc.)
+    // =============================================================================
+
+    std::unordered_map<std::string, std::string> known_mappings;
+    known_mappings["MESH"] = "MESH";
+
+    for (const auto& [mgd_type, mgd_instances] : mgd_grouping_infos) {
+        if (mgd_type == "MESH") {
+            continue;
+        }
+        for (const auto& [graph_name, _] : mgd_instances) {
+            if (!is_mgd_graph_ready(mesh_graph_descriptor, graph_name, result, known_mappings)) {
+                continue;
+            }
+            if (!mgd_graph_depends_on(mesh_graph_descriptor, graph_name, "MESH")) {
+                continue;
+            }
+            process_higher_layer_and_recurse(
+                mesh_graph_descriptor,
+                mgd_grouping_infos,
+                resolved_groupings_cache_,
+                result,
+                known_mappings,
+                mgd_type,
+                graph_name);
+        }
+    }
+
+    // Ensure all types and instances from MGD have entries in result
+    // Use MGD grouping info if no matches were found
+    for (const auto& [mgd_type, mgd_instances] : mgd_grouping_infos) {
+        for (const auto& [instance_name, mgd_grouping_info] : mgd_instances) {
+            // If not already present, use the MGD grouping info
+            if (!result[mgd_type].contains(instance_name)) {
+                result[mgd_type][instance_name].push_back(mgd_grouping_info);
+            }
+        }
+    }
+
+    return result;
+}
+
+}  // namespace tt::tt_fabric
+
+namespace {
+
+using tt::tt_metal::AsicID;
+using tt::tt_metal::ASICLocation;
+using tt::tt_metal::TrayID;
+using tt::tt_metal::experimental::tt_fabric::build_flat_adjacency_map_from_psd;
+using tt::tt_metal::experimental::tt_fabric::PhysicalAdjacencyMap;
+
+std::string build_pgd_mapping_failure_message(
+    const std::string& grouping_name,
+    const GroupingInfo& grouping_info,
+    const MappingResult<uint32_t, AsicID>& result) {
+    size_t total = grouping_info.adjacency_graph.get_nodes().size();
+    size_t mapped_count = result.target_to_global.size();
+    size_t unmapped_count = total - mapped_count;
+
+    return fmt::format(
+        "PGD grouping '{}' could not be mapped to PSD: {}/{} nodes mapped, {} unmatched",
+        grouping_name,
+        mapped_count,
+        total,
+        unmapped_count);
+}
+
+// Helper function to solve the topology mapping with pinning constraints from GroupingInfo
+MappingResult<uint32_t, AsicID> solve_for_one_grouping_to_psd(
+    const GroupingInfo& grouping_info,
+    const AdjacencyGraph<AsicID>& physical_graph,
+    const tt::tt_metal::PhysicalSystemDescriptor& physical_system_descriptor) {
+    MappingConstraints<uint32_t, AsicID> constraints;
+
+    // Build trait maps: graph nodes are 0..n-1, items[i] is the item for node i
+    std::map<uint32_t, uint32_t> target_tray_traits;
+    std::map<uint32_t, uint32_t> target_location_traits;
+
+    for (uint32_t node_id : grouping_info.adjacency_graph.get_nodes()) {
+        if (node_id >= grouping_info.items.size()) {
+            continue;
+        }
+        const GroupingItemInfo& item = grouping_info.items[node_id];
+        if (item.type != GroupingItemInfo::ItemType::ASIC_LOCATION) {
+            continue;
+        }
+        if (item.tray_id > 0) {
+            target_tray_traits[node_id] = item.tray_id;
+        }
+        if (item.asic_location > 0) {
+            target_location_traits[node_id] = item.asic_location;
+        }
+    }
+    // Build trait maps for global nodes (from physical graph)
+    std::map<AsicID, uint32_t> global_tray_traits;
+    std::map<AsicID, uint32_t> global_location_traits;
+
+    for (const auto& asic_id : physical_graph.get_nodes()) {
+        TrayID tray_id = physical_system_descriptor.get_tray_id(asic_id);
+        ASICLocation asic_location = physical_system_descriptor.get_asic_location(asic_id);
+        global_tray_traits[asic_id] = *tray_id;
+        global_location_traits[asic_id] = *asic_location;
+    }
+
+    // Add trait constraints for tray_id and asic_location
+    if (!target_tray_traits.empty() && !global_tray_traits.empty()) {
+        TT_FATAL(
+            constraints.add_required_trait_constraint<uint32_t>(target_tray_traits, global_tray_traits),
+            "Internal error: Failed to add required trait constraint for tray_id");
+    }
+    if (!target_location_traits.empty() && !global_location_traits.empty()) {
+        TT_FATAL(
+            constraints.add_required_trait_constraint<uint32_t>(target_location_traits, global_location_traits),
+            "Internal error: Failed to add required trait constraint for asic_location");
+    }
+
+    return solve_topology_mapping(
+        grouping_info.adjacency_graph, physical_graph, constraints, ConnectionValidationMode::RELAXED, true);
+}
+
+bool is_flattened(const GroupingInfo& grouping) {
+    return grouping.asic_count == grouping.adjacency_graph.get_nodes().size();
+}
+
+}  // namespace
+
+namespace tt::tt_fabric {
+
+std::unordered_set<tt::tt_metal::AsicID> PhysicalGroupingDescriptor::find_any_in_psd(
+    const GroupingInfo& grouping,
+    const tt::tt_metal::PhysicalSystemDescriptor& physical_system_descriptor,
+    std::vector<std::string>* errors_out) const {
+    // Build physical adjacency map from PSD (empty map means include all ASICs)
+    PhysicalAdjacencyMap physical_adj_map = build_flat_adjacency_map_from_psd(physical_system_descriptor, {});
+    // Convert to AdjacencyGraph
+    AdjacencyGraph<AsicID> physical_graph(physical_adj_map);
+
+    // Detect if its flattened or not, if it its not then flatten it
+    std::vector<GroupingInfo> flat_meshes;
+    if (!is_flattened(grouping)) {
+        flat_meshes = build_flattened_adjacency_mesh(grouping, physical_system_descriptor);
+    } else {
+        flat_meshes.push_back(grouping);
+    }
+
+    std::unordered_set<tt::tt_metal::AsicID> asic_ids;
+    const GroupingInfo* last_mesh_tried = nullptr;
+    MappingResult<uint32_t, AsicID> last_result;
+
+    // Use the first flat mesh that actually fits
+    for (const auto& flat_mesh : flat_meshes) {
+        if (flat_mesh.adjacency_graph.get_nodes().empty()) {
+            continue;
+        }
+
+        last_mesh_tried = &flat_mesh;
+        // solve_for_one_grouping_to_psd uses items[node_id] for trait constraints
+        auto result = solve_for_one_grouping_to_psd(flat_mesh, physical_graph, physical_system_descriptor);
+        last_result = result;
+
+        if (result.success) {
+            for (const auto& [target_node, asic_id] : result.target_to_global) {
+                asic_ids.insert(asic_id);
+            }
+            return asic_ids;
+        }
+    }
+
+    // If flat_meshes is empty, it means PSD filtering removed all possibilities
+    // This is a valid case (grouping can't be mapped to this PSD), not an internal error
+    if (flat_meshes.empty()) {
+        // Return empty set - grouping cannot be mapped to this PSD
+        return asic_ids;
+    }
+
+    // Check if there's an actual internal error (all meshes have empty graphs)
+    bool all_empty = true;
+    for (const auto& flat_mesh : flat_meshes) {
+        if (!flat_mesh.adjacency_graph.get_nodes().empty()) {
+            all_empty = false;
+            break;
+        }
+    }
+
+    if (all_empty) {
+        TT_THROW("Internal error: grouping produced empty graph");
+    }
+
+    if (errors_out != nullptr && last_mesh_tried != nullptr) {
+        errors_out->push_back(build_pgd_mapping_failure_message(grouping.name, *last_mesh_tried, last_result));
+    }
+
+    return asic_ids;
+}
+
+}  // namespace tt::tt_fabric
